initial attempt at embedding search
