+++ conf.py
@@ -25,7 +25,7 @@ import os
 import sys
 
 
-sys.path.append(os.path.abspath("."))
+sys.path.append(os.path.abspath("exts"))
 
 # -- General configuration ------------------------------------------------
 
@@ -36,12 +36,7 @@ sys.path.append(os.path.abspath("."))
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = [
-    "github",
-    "seo",
-    "components",
-    "sitemap",
-]
+extensions = ["github", "seo", "components", "sitemap", "embedding-search"]
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
