+++ exts/embedding-search/search.html
@@ -1,0 +1,207 @@
+{#
+  basic/search.html
+  ~~~~~~~~~~~~~~~~~
+
+  Template for the search page.
+
+  :copyright: Copyright 2007-2023 by the Sphinx team, see AUTHORS.
+  :license: BSD, see LICENSE for details.
+#}
+{%- extends "layout.html" %}
+{% set title = _('Search') %}
+{%- block scripts %}
+  {{ super() }}
+{%- endblock %}
+{% block extrahead %}
+<script src="{{ pathto('searchindex.js', 1) }}" defer></script>
+{{ super() }}
+{% endblock %}
+{% block body %}
+<h1 id="search-documentation">{{ _('Search') }}</h1>
+{% block scriptwarning %}
+<noscript>
+<div class="admonition warning">
+<p>
+  {% trans %}Please activate JavaScript to enable the search
+  functionality.{% endtrans %}
+</p>
+</div>
+</noscript>
+{% endblock %}
+{% block searchbox %}
+<form action="" method="get" class="search">
+  <input
+    type="text"
+    name="q"
+    aria-labelledby="search-documentation"
+    value=""
+    autocomplete="off"
+    autocorrect="off"
+    autocapitalize="off"
+    spellcheck="false"
+    placeholder="✨ Search/ask"
+  />
+  <input type="submit" value="{{ _('Go') }}" />
+</form>
+{% endblock %}
+{% block searchresults %}
+<div class="container-results">
+  <div class="loader" style="--index: 0"></div>
+  <div class="loader" style="--index: 1"></div>
+  <div class="loader" style="--index: 2"></div>
+  <div class="loader" style="--index: 3"></div>
+  <div class="loader" style="--index: 4"></div>
+  <div class="loader" style="--index: 5"></div>
+  <div class="loader" style="--index: 6"></div>
+  <div class="loader" style="--index: 7"></div>
+  <div class="loader" style="--index: 8"></div>
+  <div class="loader" style="--index: 9"></div>
+  <div id="search-results"></div>
+</div>
+{% endblock %}
+<script>
+  function cosineSimilarity(vectorA, vectorB) {
+    let dotProduct = 0;
+    for (let i = 0; i < vectorA.length; i++) {
+      dotProduct += vectorA[i] * vectorB[i];
+    }
+
+    let magnitudeA = 0;
+    let magnitudeB = 0;
+    for (let i = 0; i < vectorA.length; i++) {
+      magnitudeA += Math.pow(vectorA[i], 2);
+      magnitudeB += Math.pow(vectorB[i], 2);
+    }
+    magnitudeA = Math.sqrt(magnitudeA);
+    magnitudeB = Math.sqrt(magnitudeB);
+
+    return dotProduct / (magnitudeA * magnitudeB);
+  }
+  const articleCard = (article) => {
+    const link =
+      DOCUMENTATION_OPTIONS.URL_ROOT +
+      article.link +
+      DOCUMENTATION_OPTIONS.LINK_SUFFIX;
+    return `
+      <a href="${link}" class="article-card">
+        <p class="title">${article.title}</p>
+        <p class="link">${article.link}</p>
+      </a>
+    `;
+  };
+  const search = async () => {
+    Search._query = new URLSearchParams(window.location.search).get("q");
+    document.querySelector("input[name=q]").value = Search._query;
+
+    const resp = await fetch("https://esphome-search.ktibow.workers.dev/", {
+      method: "POST",
+      body: JSON.stringify({ query: Search._query }),
+    });
+    const data = await resp.json();
+    Search._embedding = data.data[0].embedding;
+
+    const articles = [];
+    for (const link in Search._index) {
+      const article = Search._index[link];
+      const distance = cosineSimilarity(Search._embedding, article.embedding);
+      articles.push({ link, title: article.title, distance });
+    }
+    articles.sort((a, b) => b.distance - a.distance);
+
+    const articleHtml = articles.map(articleCard).slice(0, 10).join("\n");
+    const gooogleLink =
+      "https://google.com/search?q=" +
+      encodeURIComponent(Search._query + " site:esphome.io");
+    const footerHtml = `
+      <a href="${gooogleLink}" class="search-footer" target="_blank">
+        Google it
+      </a>
+    `;
+    document.querySelector("#search-results").innerHTML =
+      articleHtml + footerHtml;
+    document.querySelector("#search-results").style.animation = "scrollIn 1s";
+  };
+  const Search = {
+    setIndex(index) {
+      Search._index = index.embeddings;
+      search();
+    },
+  };
+</script>
+<style>
+  .search {
+    display: flex;
+  }
+  .search input {
+    border: 1px solid #CCC;
+    font-family: Georgia, serif;
+    font-size: 1em;
+    padding: 0.25em;
+  }
+  .search input:first-child {
+    flex: 1 1 80%;
+  }
+  .search input:last-child {
+    flex: 1 1 20%;
+    border-left: none;
+  }
+
+  .container-results {
+    position: relative;
+    margin-top: 16px;
+  }
+  .loader {
+    background-color: #EEE;
+    height: calc(2.8em + 16px);
+    position: absolute;
+    width: 100%;
+    top: calc((2.8em + 16px + 16px) * var(--index));
+  }
+
+  #search-results {
+    position: relative;
+    z-index: 1;
+    background-color: #FFF;
+  }
+  @keyframes scrollIn {
+    0% {
+      clip-path: inset(0 0 100% 0);
+    }
+    100% {
+      clip-path: inset(0 0 calc(100% - 100vh) 0);
+    }
+  }
+
+  .article-card {
+    display: block;
+    background-color: #EEE;
+    color: #000;
+    text-decoration: none;
+    padding: 8px;
+    margin-bottom: 16px;
+  }
+  .article-card:hover {
+    text-decoration: none;
+    background-color: #DDD;
+    color: #000;
+  }
+  .article-card p {
+    margin: 0;
+  }
+  .article-card .link {
+    opacity: 0.5;
+  }
+  .search-footer {
+    display: block;
+    background-color: #22222c;
+    color: #fff;
+    padding: 8px;
+    text-decoration: none;
+  }
+  .search-footer:hover {
+    text-decoration: none;
+    background-color: #33333d;
+    color: #fff;
+  }
+</style>
+{% endblock %}
