+++ exts/embedding-search/searchbox.html
@@ -1,0 +1,22 @@
+{%- if pagename != "search" and builder != "singlehtml" %}
+<div id="searchbox" style="display: none" role="search">
+  <div class="searchformwrapper">
+    <form class="search" action="{{ pathto('search') }}" method="get">
+      <input
+        type="text"
+        name="q"
+        aria-labelledby="searchlabel"
+        autocomplete="off"
+        autocorrect="off"
+        autocapitalize="off"
+        spellcheck="false"
+        placeholder="✨ Search/ask"
+      />
+      <input type="submit" value="{{ _('Go') }}" />
+    </form>
+  </div>
+</div>
+<script>
+  document.getElementById("searchbox").style.display = "block";
+</script>
+{%- endif %}
