+++ requirements.txt
@@ -1,2 +1,3 @@
 sphinx==5.0.1
 sphinx-autobuild==2021.3.14
+python-dotenv==1.0.0
